chore: Split TelegramService into BotService, CommandService and TelegramService
